% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{algal_toxins}
\alias{algal_toxins}
\title{Algal toxins}
\format{
A data frame with 27 variables: \code{survey_year}, \code{publication_date},
\code{unique_id}, \code{site_id}, \code{date_collected}, \code{visit_number},
\code{sample_type}, \code{not_collected}, \code{visit_number}, \code{lab},
\code{analysis_material}, \code{sample_code}, \code{analyte}, \code{batch_id},
\code{sample_id}, \code{result}, \code{result_units}, \code{reporting_limit},
\code{mdl}, \code{absorbance}, \code{absorbance_units}, \code{date_received},
\code{date_analyzed}, \code{condition_code}, \code{analyst}, \code{dilution_factor},
\code{kit_id}, and \code{nars_flag}.
}
\usage{
algal_toxins
}
\description{
A dataset containing algal toxin data collected by the National Lakes Assessment.
}
\keyword{datasets}
