% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benthic_counts}
\alias{benthic_counts}
\title{Benthic counts}
\format{
A data frame with 26 variables: \code{year}, \code{unique_id}, \code{site_id},
\code{aggr_eco3_2015}, \code{aggr_eco9_2015}, \code{state}, \code{eco_bio}, \code{visit_number},
\code{date_collected}, \code{taxa_id}, \code{target_taxon}, \code{non_target},
\code{total_300}, \code{is_distinct}, \code{is_distinct_300}, \code{phylum},
\code{class}, \code{order}, \code{tribe}, \code{family}, \code{subfamily},
\code{genus}, \code{functional_feeding_group}, \code{habit},
\code{pollution_tolerance_value}, and \code{publication_date}.
}
\usage{
benthic_counts
}
\description{
A dataset containing benthic macroinvertebrate data
}
\keyword{datasets}
